Reasonable working version of Hangman!
Would like to clean up code with some serious refactoring.
There are substantial changes to be made - some classes are wrapped
instead of extended, and the input mechanism is not conducive to
running multiple tests in one command invocation.

Lastly, there are bugs in the frequency table counts which don't
seem to affect runtime correctness but are nonetheless wrong.

JUnit test harness would be a good idea to verify correctness.